make compiler happy

+++ C++/src/1.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 vector<int> Solution::twoSum(vector<int> &nums, int target) {
   unordered_map<int, int> hashMap;
-  for (int i = 0; i < nums.size(); i++) {
+  for (int i = 0; i < (int)nums.size(); i++) {
     int remainder = target - nums[i];
     if (hashMap.find(remainder) != hashMap.end()) {
       return vector<int>{hashMap[remainder], i};
